requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-jose[cryptography]
pymysql
python-dotenv
pydantic
email-validator
requests
celery
kombu